Preserve escaped sigil characters in the header
Allow characters that would otherwise be recognized as sigils to be used
as part of a field value when backslash-escaped. Lone backslashes must
also be escaped.

This is an extension to the original format that is incompatible with
existing scripts that have a bare `#` in the actor field or a `\`
anywhere.