Remove newlines immediately after override blocks
Override blocks modify the text that follows. In almost all cases where a newline would follow an override block, it is preferable to absorb the newline instead of adding a space upon removal. If a space is necessary, one can be added after the block and it will be preserved.