Remove lines containing only a comment
Comments start with %; and continue until the end of the line. This was
chosen so as to not conflict with Lua syntax. There is a risk of %;
appearing in a Lua pattern, but such instances should be trivially
replaceable with a simple ; .

Comment removal occurs as a preprocessing step and is not syntax-aware.

If the line consists of only a comment with some optional leading
whitespace, the line is removed. Doing so prevents unexpected extraneous
spaces from appearing in the output.