Compile "template" template blocks
There is no explicit check for "template" in the Effect field. Rather,
this is what the flow will be for any non-"code" template blocks.
Processing "code" blocks will be done in a follow-up commit.

On the input, the following flow is followed:

 1. Remove comments
 2. Extract blocks
 3. Process blocks
    3.1. Parse block
    3.2. Minify text
    3.3. Format as ASS event
 4. Recombine

Removing comments first allows the block delimiters to be commented out.

The same flow will be followed for "code" template blocks, but with
a different minifying function for step 3.2.