Recognize empty block headers
Given a block of the form:

    %[
      {}
    %]

recognize that the header is empty. Previously, {} was being taken by
the implementation as the Effect field.

Empty block headers were being trimmed away during block extraction,
so the parser took the first non-empty line as the header. This was not
the intention. Instead, return the entire contents of the block without
stripping. The parser already strips, so stripping during extraction is
redundant, and in this case, harmful.