Add sample aegs files
Add two aegs files that are checked by the test suite.
Both files will be referenced in the README.

The first file is format-tutorial, a walkthrough of the syntax
and features of the aegs format used by this compiler. The second file
is real-world-example, a real-world template for the ED of the LostYears
release of Fire Force Season 2, donated by The0x539, rewritten as
an aegs file.