Recognize directives during template block parsing
A directive tells aegsc to treat a template block a certain way.
Initially, this is intended for blocks that set defaults for future
blocks, but more directives may be implemented in the future.

A directive is defined as starting with an exclamation point (! U+0021)
and is separated from the Effect field by a comma (, U+002C). The
directive must be the first item in the block. It will be parsed as the
internal property "_directive".

Example of intended usage:

    %[!set-defaults, < 0:01:20.33 > 0:01:20.33 $ OP %]

The above should set the default start time, end time and style for all
following template blocks (until the next set-defaults directive) to the
supplied values.

It will be parsed as the following object:

    {
      "_dirctive": "set-defaults",
      "effect": "",
      "startTime": "0:01:20.33",
      "endTime": "0:01:20.33",
      "style": "OP",
      "text": ""
    }