Process set-defaults directives
The set-defaults directive instructs the compiler to use a new set of
defaults for all following blocks. The defaults can be set partially,
with missing fields filled in by the existing defaults.

The compiler has been refactored so that it can handle any future
directives. Each directive is associated with a processor function.
If a directive is not supported, the entire block is ignored.
If no directive is supplied, it defaults to the existing karatemplate
processor function ("karatemplate" is technically a directive, but will
not be documented).