Document !karatemplate and !set-defaults directives in format tutorial
This adds 5 new lines to the out file.

a0975c9 mentioned the karatemplate directive will not be documented,
but it is worth documenting for completeness. There could be a tool
written in the future that is simple enough that it always outputs
a directive, and knowing that the regular blocks can be represented
with a directive would allow such a tool to work.

That said, if you make a such a tool, consider taking the extra step
to omit the karatemplate directive in the output.