Strip redundant parentheses from inline code spans
With the update to the new luamin version, the tests failed because of
parentheses being preserved when they previously weren't. For inline
code spans, these parentheses are redundant because the templater
already wraps the code in a set of them before execution.

Parentheses are preserved by luamin only in some cases and stripped in
others, so outer open and close parentheses cannot be blindly stripped
without doing a full parse (e.g. `(f()) + g()` would be incorrectly
stripped to `f()) + g(` ). So instead, the expression is passed as an
argument to a function call: this allows luamin to reduce the expression
while still having known syntax that cannot be reduced wrapping it.
The function call can be stripped, leaving the reduced expression in
minified form.